issue #11:
Now indexes with complex keys should work. For some reason ServiceStack and json.net were not deserializing the cache dictionary with a complex type. This commit changes the cache key that serialized and deserialized to always be a string. There is a little overhead added to this but this overhead should only be noticeable when the index is being store (serialized) and retrieved (deserialized) from the file system. One the index is in memory, things will be the same.