issue# 14:
Now there is a PersistedDataAccess option of ReadOnly or ReadWrite that's configurable from the repository builders (both sql and file). The default is ReadWrite which means no existing behaviour has changed. Setting this value to ReadOnly means that when persisting the entities and index, if the serialized data already exists then it will not be changed. Basically, the new data will not be serialized/persisted back to file.